fix: fire and forget update tx
The old behaviour causes the exporter not to publishing prices
regularly. This change spawns a separate thread to unblock the exporter
loop. Also, because of the batch staggering behaviour to avoid calling
rpc for all batches at the same time, the old approach would always
results in waiting longer than the publishing period.