fix: remove unnecessary fresh price filtering (#103)
This change removes a check in the code that does not allow multiple
updates within the same timestamp to passthrough. `last_info.timestamp`
is always less than or equal to `info.timestamp`. So, this condition
`last_info.timestamp < info.timestamp` filters newer updates in the same
timestamp and essentially caps our publishing frequency to 1s. It was
intended to work for Solana but the publishing frequency itself takes
care of it. This change might result in increased Sol burn but the
actions to reduce it should be reducing the frequency. Please note that
in some filterings after, the same update prices are ignored so most of
the fee savings should remain there.