fix: remove adapter <> api deadlock (#108)
Adapter sends price update notifications to the subscriber
API connection actor but at the same time the subscriber
actor can get blcoked on receiving messages (e.g. product
info) from Adapter. If the channel from Adapter to the
subscriber gets full the Adapter can get blocked and
this results in a deadlock. This situation propagates
and queues messages in other actors and WS messages
from the clients will grow in an unbounded queue that
results in a constant memory growth (leak).

This change simply makes the Adaptor to subscriber API
connection nonblocking by dropping messages if the channels
are full.