chore: update Pythnet sample config
100ms notify_sched_interval is causing problems for the publishers
and this change reverts it back to 400ms.

The exact reason for the problem is still unknown but we think their
slow clients (forked from our example publisher) and poor agent design
cause this issue. A major refactor is going to improve agent bottlenecks
and also remove notifications  to make sure it doesn't add unnecessary
overhead to the clients.